Create data validation system using new DataValidator class
The DataValidator class is a generic class which keeps a list of predicates and messages. The predicates define the conditions which must be met for an object to be validated. Messages is used to provide some information about the rule that can be later be handled by the client. The Validate method returns a ValidationResult object which contains a boolean indicating success, and an array of messages relating to any conditions which haven't been met. This ValidationResult object can be handled in the client code accordingly, which is demonstrated in the AddState class where Person objects are created.