Add Id number tracking to PersonManager
Id numbers are initially set incrementally from each line in the CSV. When a new person is added, it simply continues to increment this count. That way, it will stay consistent when the contacts are saved back to CSV and loaded again. There could be a safer approach for this, but for now this will suffice.

Person validation is also now carried out in the PersonManager class rather than its clients.