Implement ability to delete contacts
Contacts can now be deleted through the delete screen, by passing a contact ID. This removes a contact from the person list if the ID exists. Also add method for getting a single person's information, which now uses an ToString override in person to get the contact information as a string.